requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
django-environ==0.11.2
environ==1.0
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
drf-yasg==1.21.7